Put image transformation into own function
