+++ src/processImage.js
@@ -4,7 +4,13 @@ import * as path from "path";
 import fsPromises from "fs/promises";
 import fs from "fs";
 import { fileTypeFromBuffer} from "file-type";
-import { extractFileName, getFileExtension, buildOutputPath, isImage} from "./processImageHelpers";
+import {
+    extractFileName,
+    getFileExtension,
+    buildOutputPath,
+    isImage,
+    transformImageWithSharp
+} from "./processImageHelpers";
 
 dotenv.config();
 
@@ -38,8 +44,12 @@ export async function saveImage(fetchResponse) {
         console.log(`Success! Your ${fileExtension} image is now copied to ${destinationFilePath}`);
         return destinationFilePath;
     } catch (error) {
-            console.error(`Error writing to ${error.path} occurred`, error.code, error.message);
-            return;
+        console.error(`Error writing to ${error.path} occurred`, error.code, error.message);
+        if (fetchResponse?.body?.cancel) {
+            console.log(`Cancelling response body`);
+            await fetchResponse.body.cancel();
+        }
+        return;
     }
 }
 
@@ -52,36 +62,10 @@ export async function processImage(inputPath) {
     const outputPath = buildOutputPath(inputPath, process.env.IMAGE_OUTPUT_PATH);
     console.log(`New output path is: ${outputPath}`);
 
-    const image = sharp(inputPath);
-    const {width, height} = await image.metadata();
-
     const borderRadius = 30;
-    const roundedCorners = Buffer.from(
-        `<svg><rect x="0" y="0" width="${width/2}" height="${height/2}" rx="${borderRadius}" ry="${borderRadius}" fill="red"/></svg>`
-    )
+    await transformImageWithSharp(inputPath, outputPath, borderRadius);
 
-    image
-        .resize(Math.round(width/2), Math.round(height/2), {
-            fit: sharp.fit.contain,
-        })
-        .composite([{
-            input: roundedCorners,
-            blend: "dest-in"
-        }])
-        .toFile(
-            outputPath, (error, info) => {
-                if (error) {
-                    console.error(`Error processing your image: ${error}. \n`);
-                    console.error(`Please try again. \n`);
-                }
-                else {
-                    console.log(`Successfully processed. \n`)
-                    console.log(`Image info = ${info.format}, width ${info.width}, heigh ${info.height}`);
-                }
-            }
-        )
     if (!fs.existsSync(outputPath)){
-        console.error(`Error: no input file at ${outputPath}`);
         return;
     }
     return outputPath;
